+++ R/Gibbs_OMIFA.R
@@ -98,7 +98,7 @@
       sigma                   <- lapply(Gseq, function(g) tcrossprod(lmat[[g]]) + diag(1/psi.inv[,g]))
       Q0                      <- Qs > 0
       log.probs               <- vapply(Gseq, function(g, Q=Q0[g]) { sigma <- if(Q) sigma[[g]] else sqrt(sigma[[g]]); dmvn(data, mu[,g], is.posi_def(sigma, make=TRUE)$X.new, log=TRUE, isChol=!Q) + log(pi.prop[g]) }, numeric(N))
-      ll.store[1]             <- sum(gumbel_max(probs=log.probs, log.like=TRUE)$log.like)
+      ll.store[1]             <- sum(rowLogSumExps(log.probs))
       Q.store[,1]             <- Qs
       G.store[1]              <- G.non
     }
@@ -133,8 +133,7 @@
       if(inherits(log.probs, "try-error")) {
         log.probs    <- vapply(Gseq, function(g, Q=Q0[g]) { sigma <- if(Q) sigma[[g]] else sqrt(sigma[[g]]); dmvn(data, mu[,g], is.posi_def(sigma, make=TRUE)$X.new, log=TRUE, isChol=!Q) + log.pis[g] }, numeric(N))
       }
-      z.res          <- gumbel_max(probs=log.probs, log.like=TRUE)
-      z              <- z.res$z
+      z              <- gumbel_max(probs=log.probs)
       nn             <- tabulate(z, nbins=G)
       nn0            <- nn  > 0
       nn.ind         <- which(nn0)
@@ -275,7 +274,7 @@
         if(sw["psi.sw"]) psi.store[,,new.it]           <- 1/psi.inv
         if(sw["pi.sw"])  pi.store[,new.it]             <- pi.prop
                          z.store[,new.it]              <- as.integer(z)
-                         ll.store[new.it]              <- if(G > 1) sum(z.res$log.like) else sum(dmvn(X=data, mu=mu[,nn.ind], sigma=tcrossprod(lmat[[nn.ind]]) + diag(psi.store[,nn.ind,new.it]), log=TRUE))
+                         ll.store[new.it]              <- sum(rowLogSumExps(log.probs))
                          Q.store[,new.it]              <- as.integer(Qs)
                          G.store[new.it]               <- as.integer(G.non)
       }